Add files via upload
predicting congestion from air quality to minimize IOT sensor costs and energy consumption. I introduced the Genetic Algorithm meta heuristic to find the right parameterization of the lstm model. The work is carried out in bi level as a type of boosting using a threshold.predicting congestion from air quality to minimize IOT sensor costs and energy consumption. I introduced the Genetic Algorithm meta heuristic to find the right parameterization of the lstm model. The work is carried out in bi level as a type of boosting using a threshold.